Replace TICKS_READ usage with get_ticks

+++ src/bird.h
@@ -39,28 +39,28 @@ typedef struct bird_s
     int slice_h;
     bird_state_t state;
     bird_color_t color_type;
-    uint32_t hit_ticks;
-    uint32_t dead_ticks;
+    uint64_t hit_ticks;
+    uint64_t dead_ticks;
     bool is_dead_reset;
     bool is_rumbling;
     bool played_die_sfx;
     int score;
     /* Animation */
-    uint32_t anim_ticks;
+    uint64_t anim_ticks;
     int anim_frame;
     /* Center point */
     float x;
     float y;
     float dx;
     float dy;
-    uint32_t dy_ticks;
+    uint64_t dy_ticks;
     /* Ready "floating" wave */
-    uint32_t sine_ticks;
+    uint64_t sine_ticks;
     float sine_x;
     float sine_y;
     /* Rotation */
     float rotation;
-    uint32_t flap_ticks;
+    uint64_t flap_ticks;
 } bird_t;
 
 /* Bird functions */
